+++ src/main/java/com/czh/modle/Comment.java
@@ -9,13 +9,13 @@ import lombok.Data;
  */
 @Data
 public class Comment {
-    private long id;
-    private long parentId;
+    private Long id;
+    private Long parentId;
     private Integer type;
-    private Integer commentator;
-    private long gmtCreate;
-    private long gmtModified;
-    private long likeCount;
+    private Long commentator;
+    private Long gmtCreate;
+    private Long gmtModified;
+    private Long likeCount;
     private String comment;
 
 }
