+++ src/main/java/com/czh/modle/Question.java
@@ -4,12 +4,12 @@ import lombok.Data;
 
 @Data
 public class Question {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
-    private long gmtCreate;
-    private long gmtModified;
-    private Integer creatorId;
+    private Long gmtCreate;
+    private Long gmtModified;
+    private Long creatorId;
     private Integer commentCount;
     private Integer viewCount;
     private Integer likeCount;
